+++ lib/fp/either/flatMap/index.ts
@@ -1,0 +1,9 @@
+import { isLeft } from ".."
+
+type FlatMapF = <E, A, U>(
+	f: (a: A) => Either<E, U>,
+) => (e: Either<E, A>) => Either<E, U>
+
+const flatMap: FlatMapF = f => e => (isLeft(e) ? e : f(e.right))
+
+export default flatMap
