+++ lib/fp/either/index.ts
@@ -1,4 +1,6 @@
-import { pipe } from "../functions"
+import ap from "./apply"
+import map from "./map"
+import flatMap from "./flatMap"
 import match from "./match"
 
 const left = <E, A = never>(e: E): Either<E, A> => ({
@@ -15,21 +17,4 @@ const isLeft = <E, A>(x: Either<E, A>): x is Left<E> => x._tag === "Left"
 
 const isRight = <A>(x: Either<any, A>): x is Right<A> => x._tag === "Right"
 
-type MapF = <E, A, U>(f: (a: A) => U) => (v: Either<E, A>) => Either<E, U>
-const map: MapF = f => v => (isLeft(v) ? v : right(f(v.right)))
-
-type FlatMapF = <E, A, U>(
-	f: (a: A) => Either<E, U>,
-) => (e: Either<E, A>) => Either<E, U>
-const flatMap: FlatMapF = f => e => (isLeft(e) ? e : f(e.right))
-
-type ApF = <E, A, U>(
-	fau: Either<E, (a: A) => U>,
-) => (e: Either<E, A>) => Either<E, U>
-const ap: ApF = fau => e =>
-	pipe(
-		fau,
-		flatMap(au => pipe(e, map(au))),
-	)
-
-export { left, right, isLeft, isRight, ap, match }
+export { left, right, isLeft, isRight, ap, match, map, flatMap }
