+++ lib/fp/option/index.ts
@@ -1,5 +1,7 @@
 import map from "./map"
+import flatMap from "./flatMap"
 import match from "./match"
+import fromNullable from "./fromNullable"
 
 const none: Option<never> = { _tag: "None" }
 
@@ -9,7 +11,4 @@ const some: SomeF = x => ({ _tag: "Some", value: x })
 type IsNoneF = <T>(x: Option<T>) => x is None
 const isNone: IsNoneF = (x): x is None => x._tag === "None"
 
-type FromNullableF = <T>(x?: T) => Option<T>
-const fromNullable: FromNullableF = x => (x == null ? none : some(x))
-
-export { none, some, isNone, fromNullable, match, map }
+export { none, some, isNone, fromNullable, match, map, flatMap }
